refactor(error): remove the redirects from the error handling
HH-309
Simplify the error page handling by removing the hackish redirects.
When an error occurs, the _error.tsx is rendered.
The _error.tsx does not fully support the getStaticPath features,
which means that it has a major issue with the translations.
The page might be rendered without the translations first,
so the translation keys are visible, but in the tests it was shown
that the translations are loaded after while.
The problem is described in https://github.com/i18next/next-i18next/issues/1020.
The 500.tsx should support the getStaticProps properly and
it is pregenerated during the build time:
https://nextjs.org/docs/pages/building-your-application/routing/custom-error#500-page.
