+++ packages/components/src/components/errorPages/ErrorPage.tsx
@@ -1,23 +1,15 @@
 import { Button, IconCrossCircle } from 'hds-react';
 import { useRouter } from 'next/router';
-import React from 'react';
-import type { Menu } from 'react-helsinki-headless-cms';
+import React, { useContext } from 'react';
 import { Page as RHHCPage } from 'react-helsinki-headless-cms';
 import { HARDCODED_LANGUAGES, MAIN_CONTENT_ID } from '../../constants';
 import useErrorsTranslation from '../../hooks/useErrorsTranslation';
 import useLocale from '../../hooks/useLocale';
+import NavigationContext from '../../navigationProvider/NavigationContext';
 import FooterSection from '../footer/Footer';
 import MatomoWrapper from '../matomoWrapper/MatomoWrapper';
 import Navigation from '../navigation/Navigation';
 import styles from './errorPage.module.scss';
-
-/** This is just a mock of footer menu to prevent
- * Apollo client to query it on error page.
- * The error page should not fetch the menu by using the Apollo client,
- * because the network error in the Apollo client is
- * usually the reason to show the error page.
- * */
-const HARDCODED_FOOTER_MENU = { menuItems: { nodes: [] } } as unknown as Menu;
 
 export type ErrorPageProps = {
   headerText: string;
@@ -33,6 +25,7 @@ const ErrorPage: React.FC<ErrorPageProps> = ({
   const { t } = useErrorsTranslation();
   const router = useRouter();
   const locale = useLocale();
+  const { footerMenu } = useContext(NavigationContext);
 
   const moveToHomePage = () => {
     router.push(`${locale}/`);
@@ -58,7 +51,11 @@ const ErrorPage: React.FC<ErrorPageProps> = ({
           </div>
         }
         footer={
-          <FooterSection menu={HARDCODED_FOOTER_MENU} appName={appName} />
+          <FooterSection
+            menu={footerMenu}
+            appName={appName}
+            hasFeedBack={false}
+          />
         }
       />
     </MatomoWrapper>
